test(e2e): retry checking subscription data (#9311)
The E2E test for publication/subscription was flaky because it checked
for replicated data immediately after creating the subscription. This
commit adds a retry mechanism (Eventually) to wait for the data to be
replicated, with a timeout of 5 minutes.

Signed-off-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
(cherry picked from commit c6f5f4801364eaa67c32321e61ab38588b8753ff)