fix: emit event and log when ImageCatalog retrieval fails (#9266)
Previously we emitted error and logs only for the errors of NotFound
category, for the other types the the controller would fail silently
without emitting a Kubernetes event. This made troubleshooting
difficult, especially when using tools like ArgoCD.

This commit ensures that a Warning event and a log is emitted when the
ImageCatalog cannot be retrieved for any reason, providing better
visibility into configuration errors.

Closes #9016

Signed-off-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Signed-off-by: Jonathan Battiato <jonathan.battiato@enterprisedb.com>
Co-authored-by: Jonathan Battiato <jonathan.battiato@enterprisedb.com>