test(e2e): skip operator unavailable test when leader election is disabled (#9310)
The "Operator unavailable" E2E test relies on scaling the operator
deployment, which behaves differently when leader election is disabled
(e.g. in GKE environments). This commit adds a check for the
LEADER_ELECTION environment variable to skip this test when leader
election is explicitly disabled. It also updates the GKE E2E workflow to
pass this environment variable to the test runner.

Closes #7229

Signed-off-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Signed-off-by: Niccolò Fei <niccolo.fei@enterprisedb.com>
Co-authored-by: Niccolò Fei <niccolo.fei@enterprisedb.com>