fix: remove Backup owner field indexer
Remove the indexer on Backup resources that allowed finding all Backups
owned by a ScheduledBackup.

The indexer never worked because it was checking for owners of kind
`Backup` instead of `ScheduledBackup`.

Even if it had correctly looked for owners of kind `ScheduledBackup`, it
would only have worked for the cases where the `ScheduledBackup` was
configured with:

```
backupOwnerReference: self
```

When `backupOwnerReference` is set to `none` (the default) or `cluster`
this indexer would not have indexed anything.

The Backup controller handles concurrency by itself by checking for
running Backups before starting a new one, so rather than fixing this indexer
we simply remove it.

Signed-off-by: Andrew Farries <andrew.farries@xata.io>