test(configuration-e2e): skip simultaneous update when running on a development version of Postgres (#9399)
Avoid running the simultaneous update test when we are testing a Pg
major that's greater then the `DefaultImageName` specified by the
operator.

The `DefaultImageName` always points to the latest stable major
version, thus if the image we are testing is a newer Pg Major, that
most likely means we are running on a development version of PG (like
the ones being built in
[postgres-trunk-containers](https://github.com/cloudnative-pg/postgres-trunk-containers/)).

The simultaneous update test requires applying an extension, which for
development versions of PG should be built from source and we have no
guarantee that the extension is compatible yet.

Closes #9331

Signed-off-by: Niccolò Fei <niccolo.fei@enterprisedb.com>
Signed-off-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>
Co-authored-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>
(cherry picked from commit 83a7c72ff9def5aa5608c840eb1358e8447d67bc)