fix(backup): use correct pod name in event when pod is not found
In isValidBackupRunning, when the elected backup pod is not found, the
event message was incorrectly referencing pod.Name which is empty since
the Get() call returned NotFound. Changed to use the actual pod name
from backup.Status.InstanceID.PodName which contains the name of the
pod that couldn't be found.

Signed-off-by: Armando Ruocco <armando.ruocco@enterprisedb.com>